requirements.txt
# Core ML dependencies
torch>=1.12.0
torchvision>=0.13.0
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.12.0

# Image processing
Pillow>=8.3.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Utilities
tqdm>=4.64.0
matplotlib>=3.5.0

# Optional but recommended for better performance
xformers>=0.0.16
